Add an early-out check in IsMemoryReadable
In IsMemoryReadable, if the address is small enough, then return early
rather than messing around with a pipe.

We'll generally hit this case for the version field at the top of
wl_egl_window, since the version number will be a small integer.

This is partly an optimization, but it also avoids triggering an error
report from valgrind in an otherwise correct program.