Add support for PRIME.
Most of the logic that we needed was already implemented in the various
support code, so this just finishes up what was left in eglInitialize
and eglSwapBuffers.

Added a WlDisplayInstance::force_prime flag to indicate that the surface
code should use PRIME even if it looks like there's a supported
modifier.

In eglInitialize, allow the case where the render device is not the
server device. Added code to open the correct DRM node for the render
device, rather than the server device.

In PickDefaultModifiers, if force_prime is set, then leave the surface's
modifier list empty. That tells SwapChainRealloc to allocate a PRIME
swapchain, using only the driver's modifier list.

In eplWlSwapBuffers, if PRIME is enabled, then it'll look for a free
buffer first and blit to it. After presentation, we only assign a new
back buffer for non-PRIME.