Implement per-surface dma-buf feedback.
Add a SurfaceFeedbackState struct to keep track of a per-surface
zwp_linux_dmabuf_feedback_v1 object.

It uses WlDmaBufFeedbackCommon to share the common stuff with the
default feedback handling.

The rest of the bookkeeping is relatively simple: All we really need is
a bool array for which of the driver's format modifiers is supported. We
can ignore any fourcc formats that don't match the surface, and we can
ignore any tranches that don't match the render device.

When we get a zwp_linux_dmabuf_feedback_v1::done event, it'll use that
bool array to fill in the existing modifier array, so the swapchain
allocation code doesn't have to care where the modifier list came from.