Use a separate wl_event_queue for each WlSwapChain.
Create a separate wl_event_queue to handle events for each WlSwapChain.

The only thing that we actually use the surface's event queue for is
waiting for wl_buffer::release events, and that only happens when we're
not using explicit sync.

Using a separate event queue for the swapchain ensures that
eplWlSwapChainFindFreePresentBuffer behaves the same way regardless of
whether it's using explicit sync or not.