Add an environment variable to override protocol versions.
If __NV_WAYLAND_PROTOCOL_VERSIONS is set, then it's parsed as a
comma-separated list of "PROTOCOL=version" strings, which can be used to
either disable a protocol or to limit what version the library will try
to use.

This is mostly just for testing, but it might also be useful as a
workaround if there's some bug with a protocol that the platform library
can optionally use.