Use the FIFO and commit timing protocls in eglSwapBuffers
If wp_commit_timer_v1, wp_fifo_manager_v1, and wp_presentation_feedback
are all available, then use them to implement frame throttling for
eglSwapBuffers.

Between the three of them (and an ugly workaround with a dummy commit),
it's possible to implement frame throttling without causing
eglSwapBuffers to hang when the window is occluded.