Rework swap interval handling in eglSwapBuffers.
In eglSwapBuffers, only wait for previous frames if the swap interval is
positive. If the swap interval is zero, then we should try to present as
soon as possible.

eglSwapBuffers will now unconditionally request presentation feedback or
a frame callback, though whether it waits for either will depend on the
what the swap interval is in the next eglSWapBuffers.