In eglSwapBuffers, always send wp_fifo_v1::set_barrier
In eglSwapBuffers, if we have a wp_fifo_v1 object, then always send a
set_barrer request, no matter what the current swap interval is.

If the swap interval is zero, then we don't send a wait_barrier request,
but it means that if the swap interval is set to nonzero after the swap,
then the next eglSwapBuffers can wait on it.