Fix presentation feedback bookkeeping for eglWaitGL.
We can't safely use presentation feedback in eglWaitGL if the swap
interval was zero.

If the swap interval was zero, then we can't use
wp_fifo_v1::wait_barrier, which means that we can't use the
double-commit workaround, and so presentation feedback may be delayed
indefinately.

Instead, only request presentation feedback when the swap interval is
nonzero, but send a wl_display::sync request at the very end. That way,
eglWaitGL can at least make sure that the server has received the commit
requset.