Fall back to gbm_bo_create_with_modifiers.
If we're running with an older verison of libgbm that doesn't have
gbm_bo_create_with_modifiers2, then fall back to using
gbm_bo_create_with_modifiers instead.

To check for whether libgbm is new enough, check if gbm_bo_get_bpp is
available. We don't use that function, but it's the first symbol added
after GBM_BO_IMPORT_FD_MODIFIER, which we do require.

Change the README to specify libgbm 21.2 as a requirement, to match the
egl-gbm platform library.