Implement EGL_EXT_present_opaque.
Add a present_fourcc member to EplImplSurface, which gets used for the
fourcc code that we send to the server and for parsing the server's
dma-buf feedback.

Change eplWlCreateWindowSurface so that it will parse the attribute list
and remove an EGL_PRESENT_OPAQUE_EXT attribute, if present.

If the EGL_PRESENT_OPAQUE_EXT attribute is true, then it will convert
the fourcc code from the EGLConfig to a corresponding opaque format, and
assign that to EplImplSurface::present_fourcc.

Also update the asserts in PickDefaultModifiers, since it is possible
that the server could advertise different support for
otherwise-equivalent opaque and transparent formats, and checking for
EGL_WINDOW_BIT only checks the original (transparent) format.