Don't pass EGL_RENDER_BUFFER to the driver.
In eplWlCreateWindowSurface, if the application passes the
EGL_RENDER_BUFFER attribute, then consume it locally instead of passing
it to the driver.

If the application requests EGL_SINGLE_BUFFER, then we'll issue a
warning, but not an error. Wayland doesn't allow front-buffered
rendering, but that attribute is only a hint.