Replace SurfaceFeedbackState::modifiers_changed with a counter
Currently, we set the modifiers_changed flag upon receiving new surface
feedback data, but we never clear it, and there isn't a good place to do
so.

We shouldn't clear the flag after calling SwapChainRealloc in
SwapBuffers, because we haven't set the new swapchain yet, and we might
hit an error before we actually call SetWindowSwapchain.

We also shouldn't clear the flag when we call SetWindowSwapchain,
because it's possible that another round of feedback data could arrive
between SwapChainRealloc and SetWindowSwapchain.

Instead, replace the modifiers_changed flag with a counter that we
increment when we get a zwp_linux_dmabuf_feedback_v1::done event, and
we record the value of that counter in WlSwapChain. If the swapchain's
counter doesn't match the window's counter, then we know we need to scan
the modifier list again.