Don't use the wl_display during teardown.
Some applications will call wl_display_disconnect but not eglTerminate
before exiting, which leads to segfaults or hangs when we try to clean
up the EGLDisplay.

While that's technically an app bug, it's a common enough one to be
worth trying to work around.

To do that, add a new eplWlDisplayInstanceIsNativeValid function, which
returns true if we expect the wl_display to still be valid. In this
case, that means either we're not in teardown, or the native display was
NULL and so egl-wayland2 has its own wl_display. We can update that if
Wayland ever gains an equivalent to XESetCloseDisplay.

In eplWlDestroyWindow and eplWlDisplayInstanceFree, avoid calling any
Wayland functions if eplWlDisplayInstanceIsNativeValid returns false.
We'll just leak any queues or proxy structs that libwayland didn't clean
up.